04.11.2022 prod changes commit change event and sale event -> field weight from integer to double

+++ src/main/java/uz/sngroup/model/event/Event.java
@@ -41,7 +41,7 @@ public class Event {
 
     private Integer pieces = 0;
 
-    private Integer weight = 0;
+    private Double weight = 0.0;
 
     private String description = " ";
